added static cast to avoid long long issue

+++ src/tests/numbers/Test_GMP.cpp
@@ -4,7 +4,7 @@
 TEST(GMP, Debug)
 {
 	carl::sint i = 4;
-	mpq_class n(i);
+	mpq_class n(static_cast<signed long>(i));
 }
 
 TEST(GMP, Rationalize)
